requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
pytest-asyncio==0.21.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.23.2
zipp==3.17.0
